requirements.txt
Flask
flask-pymongo
flask-bcrypt
flask-jwt-extended
python-dotenv
Flask-Cors
dnspython